Agregar balance al usuario

+++ db/migrate/20160703054905_add_balance_to_usuarios.rb
@@ -1,5 +1,5 @@
 class AddBalanceToUsuarios < ActiveRecord::Migration
   def change
-    add_column :usuarios, :balance, :integer
+    add_column :usuarios, :balance, :decimal
   end
 end
